Reorganize fuzz infrastructure to facilitate running from outside
