+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
     "soroban-builtin-sdk-macros",
     "soroban-test-wasms",
     "soroban-synth-wasm",
+    "soroban-fuzz-targets",
     "soroban-bench-utils",
     "soroban-simulation",
 ]
