+++ soroban-env-host/Cargo.toml
@@ -77,7 +77,7 @@ serde = "1.0.192"
 arbitrary = "1.3.2"
 lstsq = "0.5.0"
 nalgebra = { version = "0.32.3", default-features = false, features = ["std"]}
-wasm-encoder = "0.36.2"
+wasm-encoder = "0.38.1"
 rustversion = "1.0"
 wycheproof = "0.5.1"
 k256 = {version = "0.13.1", default-features = false, features = ["alloc"]}
