+++ soroban-fuzz-targets/src/lib.rs
@@ -1,0 +1,27 @@
+//! Fuzz target entry points for external harnesses.
+//!
+//! This module provides functions that implement the core logic of fuzz
+//! targets, callable from external harnesses like stellar-core or cargo-fuzz.
+//!
+//! The functions take raw bytes as input (as provided by libfuzzer/AFL/etc) and
+//! return a result indicating whether the input was valid and what happened.
+//!
+//! The actual fuzz target entry points that call these functions live in a
+//! separate workspace (soroban-env-host/fuzz) to avoid pulling in fuzzing
+//! dependencies into the main workspace and to allow the fuzz target bodies in
+//! this crate to be called from multiple different fuzzing harnesses, including
+//! those in external workspaces like stellar-core.
+
+pub mod expr;
+pub mod wasmi;
+
+/// Result of running a fuzz target.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum FuzzResult {
+    /// Input was successfully processed (contract may have returned error, but no internal errors)
+    Ok,
+    /// Input was rejected (too short, malformed, etc) - not a bug
+    Reject,
+    /// An internal error occurred - this is a bug we want to find
+    InternalError,
+}
