+++ soroban-synth-wasm/Cargo.toml
@@ -11,7 +11,7 @@ rust-version.workspace = true
 publish = false
 
 [dependencies]
-wasm-encoder = "=0.36.2"
+wasm-encoder = "=0.38.1"
 wasmparser = { workspace = true }
 soroban-env-common = { workspace = true }
 stellar-xdr = { workspace = true, features = ["std"] }
